+++ client/src/main/java/net/oneandone/stool/client/cli/Ssh.java
@@ -15,45 +15,33 @@
  */
 package net.oneandone.stool.client.cli;
 
+import io.fabric8.kubernetes.client.dsl.ExecWatch;
 import net.oneandone.stool.client.Globals;
+import net.oneandone.stool.client.OpenShift;
+import net.oneandone.stool.client.PodConfig;
 import net.oneandone.stool.client.Reference;
-import net.oneandone.sushi.fs.file.FileNode;
-import net.oneandone.sushi.launcher.Launcher;
 
 import java.io.IOException;
 
 public class Ssh extends IteratedStageCommand {
-    public Ssh(Globals globals) {
+    private final int timeout;
+
+    public Ssh(Globals globals, int timeout) {
         super(globals);
+        this.timeout = timeout;
     }
 
     @Override
     public void doMain(Reference reference) throws IOException, InterruptedException {
-        String str;
-        Launcher launcher;
-        int result;
-        FileNode privateKey;
+        PodConfig config;
 
-        str = reference.client.ssh(reference.stage);
-        privateKey = world.getTemp().createTempFile();
-        privateKey.setPermissions("rwx------");
-        privateKey.writeString(str);
-        try {
-            launcher = world.getWorking().launcher("ssh");
-            launcher.arg("stool@" + reference.client.getServer());
-            if (console.getVerbose()) {
-                launcher.arg("-v");
+        config = reference.client.podToken(reference.stage, timeout);
+        console.verbose.println(config.toString());
+        try (OpenShift os = OpenShift.create(config)) {
+            try (ExecWatch watch = os.ssh(config.pod)) {
+                Thread.sleep(timeout * 60 * 1000);
+                console.info.println("timed out - terminating");
             }
-            launcher.arg("-i");
-            launcher.arg(privateKey.getAbsolute());
-
-            launcher.env("SSH_AUTH_SOCK", ""); // make sure not to save keys: disable agent
-            console.info.println("starting ssh shell to " +  reference.stage);
-            launcher.getBuilder().inheritIO();
-            result = launcher.getBuilder().start().waitFor();
-            console.verbose.println("result: " + result);
-        } finally {
-            privateKey.deleteFile();
         }
     }
 }